Fix: store color and mouse position in seedling
